+++ lib/src/widget.dart
@@ -2,10 +2,11 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'constant.dart';
-
 const _list = [0, 23, 255];
 const _bullet = [8226, 9702, 9642];
+const um = 20;
+const double inch = 2.54;
+const dot = 72.0;
 
 enum _Status { starting, normal, exit }
 
@@ -69,18 +70,12 @@ class PageFormat {
         marginLeft = marginAll ?? marginLeft,
         marginRight = marginAll ?? marginRight;
 
-  static const PageFormat a3 =
-      PageFormat(29.7 * cm, 42 * cm, marginAll: 2.0 * cm);
-  static const PageFormat a4 =
-      PageFormat(21.0 * cm, 29.7 * cm, marginAll: 2.0 * cm);
-  static const PageFormat a5 =
-      PageFormat(14.8 * cm, 21.0 * cm, marginAll: 2.0 * cm);
-  static const PageFormat a6 =
-      PageFormat(10.5 * cm, 14.8 * cm, marginAll: 1.0 * cm);
-  static const PageFormat letter =
-      PageFormat(8.5 * dot, 11.0 * dot, marginAll: dot);
-  static const PageFormat legal =
-      PageFormat(8.5 * dot, 14.0 * dot, marginAll: dot);
+  static const PageFormat a3 = PageFormat(29.7 * cm, 42 * cm, marginAll: 2.0 * cm);
+  static const PageFormat a4 = PageFormat(21.0 * cm, 29.7 * cm, marginAll: 2.0 * cm);
+  static const PageFormat a5 = PageFormat(14.8 * cm, 21.0 * cm, marginAll: 2.0 * cm);
+  static const PageFormat a6 = PageFormat(10.5 * cm, 14.8 * cm, marginAll: 1.0 * cm);
+  static const PageFormat letter = PageFormat(8.5 * dot, 11.0 * dot, marginAll: dot);
+  static const PageFormat legal = PageFormat(8.5 * dot, 14.0 * dot, marginAll: dot);
 
   PageFormat copyWith(
       {double? width,
@@ -122,6 +117,8 @@ class PageFormat {
 
 /// RTF Document
 class Document {
+  late IOSink _out;
+
   /// the widgets list
   final List<Widget> _elements;
   final List<_Font> _fonts = [];
@@ -136,10 +133,15 @@ class Document {
       {PageFormat pageFormat = PageFormat.a4,
       int orientation = 1,
       bool interleave = false,
-      Widget? hdLeft, Widget? hdCenter, Widget? hdRight, Widget? ftLeft, Widget? ftCenter, Widget? ftRight})
+      Widget? hdLeft,
+      Widget? hdCenter,
+      Widget? hdRight,
+      Widget? ftLeft,
+      Widget? ftCenter,
+      Widget? ftRight})
       : _pageFormat = pageFormat,
         _orientation = orientation,
-        _interleave = interleave, 
+        _interleave = interleave,
         _hfValues = [hdLeft, hdCenter, hdRight, ftLeft, ftCenter, ftRight];
   void _newWidget() {
     if (_status == _Status.exit) {
@@ -150,39 +152,39 @@ class Document {
   }
 
   void _newSection(int nCols, int spCol) {
-    if (_status != _Status.starting) write('\\sect');
+    if (_status != _Status.starting) _out.write('\\sect');
     _status = _Status.normal;
     var l = um * spCol;
-    write('\\cols$nCols\\colsx$l\r\n');
+    _out.writeln('\\cols$nCols\\colsx$l');
   }
 
   Future<void> save(File f, {int charset = 1252, int lang = 1040}) async {
-    out = f.openWrite(encoding: latin1);
-    write("{\\rtf1\\ansi\\ansicpg");
-    write(charset.toString());
-    write("\\uc1 \\deff0\\deflang$lang");
-    write("\\deflangfe$lang");
-    write("{\\fonttbl");
+    _out = f.openWrite(encoding: latin1);
+    _out.write("{\\rtf1\\ansi\\ansicpg");
+    _out.write(charset.toString());
+    _out.write("\\uc1 \\deff0\\deflang$lang");
+    _out.write("\\deflangfe$lang");
+    _out.write("{\\fonttbl");
     _defineFonts();
-    write("}");
+    _out.write("}");
 
     _defineColors();
 
-    write("{\\stylesheet");
+    _out.write("{\\stylesheet");
     for (int index = 0; index < _fonts.length; index++) {
       _defineStyle(_fonts[index], index, lang);
     }
-    write("}");
+    _out.write("}");
     _defineListTables();
     _defineHeader();
 
     _hf(false);
     _hf(true);
     for (var r in _elements) {
-      r.draw(this);
+      r.draw(this, _out);
     }
-    write("\\pard}\r\n");
-    await out.close();
+    _out.writeln("\\pard}");
+    await _out.close();
   }
 
   /// returns the current PageFormat
@@ -195,9 +197,9 @@ class Document {
   }
 
   /// add a new Font to the fonts list
-  void addFont(String name, String family, String fontName, FontStyle style,
-          double size) =>
+  void addFont(String name, String family, String fontName, FontStyle style, double size) =>
       _fonts.add(_Font(name, family, fontName, style, size));
+
   /// add an element to the header or footer
   /// Warning: if an Image widget will be added and its height is too big, an invalid document could be generated
   @Deprecated('use hdLeft, hdCenter, hdRight, ftLeft, ftCenter, ftRight in Document constructor')
@@ -268,55 +270,54 @@ class Document {
   }
 
   void _defineListTables() {
-    write('{\\*\\listtable');
+    _out.write('{\\*\\listtable');
     for (int i = 1; i <= _list.length; i++) {
-      write('{\\list\\listtemplateid$i\n');
+      _out.writeln('{\\list\\listtemplateid$i');
       for (int lvl = 0; lvl < 8; lvl++) {
-        write(
+        _out.write(
             '{\\listlevel\\levelnfc${_list[i - 1]}\\leveljc0\\levelstartat1\\levelfollow${i == 3 ? 2 : 0}{\\leveltext ');
-        //if (i == 1) {
         switch (i) {
           case 1:
-            write('\\\'02\\\'0$lvl.;}');
+            _out.write('\\\'02\\\'0$lvl.;}');
             break;
           case 2:
-            write('\\\'01\\u${_bullet[lvl % _bullet.length]};}');
+            _out.write('\\\'01\\u${_bullet[lvl % _bullet.length]};}');
             break;
           case 3:
-            write('\\\'00;}');
+            _out.write('\\\'00;}');
             break;
         }
-        write('{\\levelnumbers${i == 1 ? '\\\'01' : ''};}');
+        _out.write('{\\levelnumbers${i == 1 ? '\\\'01' : ''};}');
         if (i == _list.length) {
-          write('\\fi0\\li0}\n');
+          _out.writeln('\\fi0\\li0}');
         } else {
-          write('\\fi-360\\li${720 + 360 * lvl}}\n');
+          _out.writeln('\\fi-360\\li${720 + 360 * lvl}}');
         }
       }
-      write('\\listid$i}');
+      _out.write('\\listid$i}');
     }
-    write('}');
-    write('{\\listoverridetable');
+    _out.write('}');
+    _out.write('{\\listoverridetable');
     for (int i = 1; i <= _list.length; i++) {
-      write('{\\listoverride\\listid$i\\listoverridecount0\\ls$i}');
+      _out.write('{\\listoverride\\listid$i\\listoverridecount0\\ls$i}');
     }
-    write('}');
+    _out.writeln('}');
   }
 
   void _defineFonts() {
     for (int idx = 0; idx < _fonts.length; idx++) {
-      write("{\\f$idx");
-      write("\\f${_fonts[idx]._family}");
-      write("\\fcharset0\\fprq2 ${_fonts[idx]._fontName};}");
+      _out.write("{\\f$idx");
+      _out.write("\\f${_fonts[idx]._family}");
+      _out.write("\\fcharset0\\fprq2 ${_fonts[idx]._fontName};}");
     }
   }
 
   void _defineColors() {
-    write("{\\colortbl;");
+    _out.write("{\\colortbl;");
     for (Color bc in Color.values) {
-      write('\\red${bc._red()}\\green${bc._green()}\\blue${bc._blue()};');
+      _out.write('\\red${bc._red()}\\green${bc._green()}\\blue${bc._blue()};');
     }
-    write("\\red255\\green255\\blue255;}\r\n");
+    _out.writeln("\\red255\\green255\\blue255;}");
   }
 
   void _defineStyle(_Font f, int index, int lang) {
@@ -342,26 +343,26 @@ class Document {
       str.write('\\sbasedon0 ');
     }
     str.write('\\snext0 $styleName;}');
-    write('{$str');
+    _out.write('{$str');
   }
 
   void _defineHeader() {
     var page = getPage();
-    write(
-        "\\paperw${(um * page.width).round()}\\paperh${(um * page.height).round()}\\margl${(um * page.marginLeft).round()}\\margr${(um * page.marginRight).round()}\\margt${(um * page.marginTop).round()}\\margb${(um * page.marginBottom).round()}${_orientation == 0 ? "\\landscape" : ""} \n\\widowctrl\\ftnbj\\aenddoc\\hyphcaps0\\viewkind1\\viewscale90");
-    write("\\fet0");
-    write("\\sectd ");
-    write("\\linex0");
-    write("\\headery500");
-    if(_hfValues.sublist(3).where((element) => element != null).isNotEmpty) {
-      write("\\footery600");
+    _out.writeln(
+        "\\paperw${(um * page.width).round()}\\paperh${(um * page.height).round()}\\margl${(um * page.marginLeft).round()}\\margr${(um * page.marginRight).round()}\\margt${(um * page.marginTop).round()}\\margb${(um * page.marginBottom).round()}${_orientation == 0 ? "\\landscape" : ""}");
+    _out.write("\\widowctrl\\ftnbj\\aenddoc\\hyphcaps0\\viewkind1\\viewscale90");
+    _out.write("\\fet0\\sectd ");
+    _out.write("\\linex0");
+    _out.write("\\headery500");
+    if (_hfValues.sublist(3).where((element) => element != null).isNotEmpty) {
+      _out.write("\\footery600");
     }
-    write("\\sectdefaultcl \\pard\\plain ");
+    _out.write("\\sectdefaultcl \\pard\\plain ");
   }
 
   _Font _setStyle(String font) {
     int idx = _fonts.indexWhere((e) => e._name == font);
-    write('\\pard\\plain${_mStyles[idx]}');
+    _out.write('\\pard\\plain${_mStyles[idx]}');
     return _fonts[idx];
   }
 
@@ -371,39 +372,37 @@ class Document {
     if (!styles.any((s) => s != null)) {
       return;
     }
-    write(bFooter ? "{\\footer " : "{\\header ");
+    _out.write(bFooter ? "{\\footer " : "{\\header ");
     // don't round before multiply, otherwise you loose some decimals
-    write(
-        "\\pard\\plain \\nowidctlpar\\widctlpar\\tqc\\tx${l.round()}\\tqr\\tx${(2 * l).round()}\\adjustright {");
-    write("\\b\\qj ");
+    _out.write("\\pard\\plain \\nowidctlpar\\widctlpar\\tqc\\tx${l.round()}\\tqr\\tx${(2 * l).round()}\\adjustright {");
+    _out.write("\\b\\qj ");
     for (int j = 0; j < styles.length; j++) {
       if (j > 0) {
-        write("\\tab ");
+        _out.write("\\tab ");
       }
       var s = styles[j];
       if (s != null) {
-        s.draw(this);
+        s.draw(this, _out);
       }
     }
-    write("\\par }}");
+    _out.write("\\par }}");
   }
 
-  void _insertImage(
-      ByteData data, int dpi, bool share, int? width, int? height) {
+  void _insertImage(ByteData data, int dpi, bool share, int? width, int? height) {
     if (share) {
-      write("{");
-      write("\\*\\shppict ");
+      _out.write("{");
+      _out.write("\\*\\shppict ");
     }
-    write("{");
-    write("\\pict ");
-    write("\\pngblip ");
+    _out.write("{");
+    _out.write("\\pict ");
+    _out.write("\\pngblip ");
 
     int mult = 1440 ~/ dpi;
     if (width != null) {
-      write("\\picwgoal ${(width * mult)} ");
+      _out.write("\\picwgoal ${(width * mult)} ");
     }
     if (height != null) {
-      write("\\pichgoal${(height * mult)} ");
+      _out.write("\\pichgoal${(height * mult)} ");
     }
     for (int i = 0; i < data.buffer.lengthInBytes; i++) {
       int iData = data.getInt8(i);
@@ -414,13 +413,13 @@ class Document {
       }
       if (iData < 16) {
         // Set leading zero and append
-        write("0");
+        _out.write("0");
       }
-      write(iData.toRadixString(16));
+      _out.write(iData.toRadixString(16));
     }
 
-    write("}");
-    write("}");
+    _out.write("}");
+    _out.write("}");
   }
 }
 
@@ -434,24 +433,23 @@ class _Font {
   final double _size;
   final String _family;
 
-  const _Font(
-      this._name, this._family, this._fontName, this._style, this._size);
+  const _Font(this._name, this._family, this._fontName, this._style, this._size);
 }
 
 /// the Widget class
 abstract class Widget {
-  void draw(Document doc);
+  void draw(Document doc, IOSink out);
+  int col() => 1;
 }
 
-/*
-abstract class _SingleChildWidget extends Widget {
+abstract class SingleChildWidget extends Widget {
   Widget child;
-  _SingleChildWidget({required this.child});
+  SingleChildWidget({required this.child});
   @override
-  void draw(Document doc) {
-    child.draw(doc);
+  void draw(Document doc, IOSink out) {
+    child.draw(doc, out);
   }
-}*/
+}
 
 abstract class _MultipleChildrenWidget extends Widget {
   /// the children widgets list
@@ -482,8 +480,8 @@ class TextStyle {
 /// Widget to insert a new page break
 class SkipPage extends Widget {
   @override
-  void draw(Document doc) {
-    write("\\pard\\plain \\page \r\n");
+  void draw(Document doc, IOSink out) {
+    out.writeln("\\pard\\plain \\page");
   }
 }
 
@@ -493,12 +491,12 @@ class PageNo extends Widget {
   final bool nofPages;
   PageNo({this.nofPages = true});
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     if (nofPages) {
-      write(
+      out.write(
           "{\\field{\\*\\fldinst { PAGE }}{\\fldrslt { 1}}}/{\\field{\\*\\fldinst { NUMPAGES    \\* MERGEFORMAT }}{\\fldrslt { 1}}}");
     } else {
-      write("{\\field{\\*\\fldinst { PAGE }}{\\fldrslt { 1}}}");
+      out.write("{\\field{\\*\\fldinst { PAGE }}{\\fldrslt { 1}}}");
     }
   }
 }
@@ -506,20 +504,24 @@ class PageNo extends Widget {
 /// Widget to insert a new line break
 class NewLine extends Widget {
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     doc._newWidget();
-    write("\\par ");
+    out.write("\\par ");
   }
 }
 
 /// this widget show all its children separated by a newline
 class Column extends _MultipleChildrenWidget {
-  Column({required super.children});
+  /// if false, omit newLine after last child
+  final bool lastNL;
+  Column({required super.children, this.lastNL = true});
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     for (var c in children) {
-      c.draw(doc);
-      write('\\par ');
+      c.draw(doc, out);
+      if (lastNL || c != children.last) {
+        out.write('\\par ');
+      }
     }
   }
 }
@@ -528,9 +530,9 @@ class Column extends _MultipleChildrenWidget {
 class Row extends _MultipleChildrenWidget {
   Row({required super.children});
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     for (var c in children) {
-      c.draw(doc);
+      c.draw(doc, out);
     }
   }
 }
@@ -539,13 +541,12 @@ class Row extends _MultipleChildrenWidget {
 class Section extends Column {
   final int _nCols;
   final int _spCol;
-  Section(this._nCols, {required super.children, int? spCol})
-      : _spCol = spCol ?? 36;
+  Section(this._nCols, {required super.children, int? spCol}) : _spCol = spCol ?? 36;
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     doc._sections.add(this);
     doc._newSection(_nCols, _spCol);
-    super.draw(doc);
+    super.draw(doc, out);
     doc._sections.removeLast();
     doc._status = _Status.exit;
   }
@@ -558,14 +559,14 @@ class Text extends Widget {
   Text(this._txt, {TextStyle? style}) : _style = style;
 
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     doc._newWidget();
     String a = '';
     if (_style != null) {
       doc._setStyle(_style._style);
       a = _style._align?._al ?? '';
     }
-    write('{$a${doc._text(_txt)}}');
+    out.write('{$a${doc._text(_txt)}}');
   }
 }
 
@@ -585,10 +586,9 @@ class Image extends Widget {
 
   /// image's height
   final int? height;
-  Image(this._data,
-      {this.dpi = 72, this.share = true, this.width, this.height});
+  Image(this._data, {this.dpi = 72, this.share = true, this.width, this.height});
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     doc._newWidget();
     doc._insertImage(_data, dpi, share, width, height);
   }
@@ -600,12 +600,12 @@ class Line extends Widget {
   final double w;
   Line([this.w = 1]);
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     doc._newWidget();
     double width = doc.getPage().availableWidth;
     double wd = w * width;
     double mg = (um * width - wd) / 2;
-    write(
+    out.write(
         "\\pard\\plain {\\shp{\\*\\shpinst\\shpleft${mg.round()}\\shptop180\\shpright${(mg + wd).round()}\\shpbottom180\\shpfhdr0\\shpbxcolumn\\shpbxignore\\shpbypara\\shpbyignore\\shpwr3\\shpwrk0\\shpfblwtxt0\\shpz0\\shplid1028{\\sp{\\sn shapeType}{\\sv 20}}{\\sp{\\sn fFlipH}{\\sv 0}}{\\sp{\\sn fFlipV}{\\sv 0}}{\\sp{\\sn shapePath}{\\sv 4}}{\\sp{\\sn fFillOK}{\\sv 0}}{\\sp{\\sn fFilled}{\\sv 0}}{\\sp{\\sn fArrowheadsOK}{\\sv 1}}{\\sp{\\sn fLayoutInCell}{\\sv 1}}}{\\shprslt{\\*\\do\\dobxcolumn\\dobypara\\dodhgt8192\\dpline\\dpptx0\\dppty0\\dpptx9900\\dppty0\\dpx0\\dpy0\\dpxsize\\dpysize0\\dplinew15\\dplinecor0\\dplinecog0\\dplinecob0}}}");
   }
 }
@@ -616,13 +616,13 @@ class Listing extends Widget {
   final bool _numbered;
   Listing(this._elements, this._numbered);
   @override
-  void draw(Document doc) {
+  void draw(Document doc, IOSink out) {
     doc._newWidget();
     for (int i = 0; i < _elements.length; i++) {
       String c = _numbered ? '$i.)' : '\\u8226\\\'95';
-      write('{\\listtext\\pard\\plain $c\\tab}\\ilvl0\\ls${_numbered ? 1 : 2}');
-      _elements[i].draw(doc);
-      write('\\par');
+      out.write('{\\listtext\\pard\\plain $c\\tab}\\ilvl0\\ls${_numbered ? 1 : 2}');
+      _elements[i].draw(doc, out);
+      out.write('\\par');
     }
   }
 }