+++ CHANGELOG.md
@@ -4,11 +4,17 @@
 
 ## 0.0.2
 
-* Added a Line Widget that show a line
-* Added a Listing Widget that show a list
+* Added a Line widget that show a line
+* Added a Listing widget that show a list
 
 ## 0.0.3
 
 * Added the Section widget that show its children in one or more columns
 * Changed the way to define document header/footer
 * Added some examples
+
+## 0.0.4
+
+* Reengineered code
+* Added a parameter to Column widget to omit newLine after last child
+* Added a ColSpan widget to span a cell on more columns